requirements.txt
psycopg2-binary==2.8.*
beautifulsoup4
aiohttp
lxml